fix: use build flags for os dependent code
