+++ pkg/process/checks/process_nix.go
@@ -8,21 +8,24 @@
 package checks
 
 import (
+	"fmt"
 	"math"
 	"os/user"
+	"slices"
 	"strconv"
 
 	model "github.com/DataDog/agent-payload/v5/process"
 	"github.com/shirou/gopsutil/v4/cpu"
 
+	"github.com/DataDog/datadog-agent/comp/core/tagger/tags"
 	"github.com/DataDog/datadog-agent/pkg/process/procutil"
+	"github.com/DataDog/datadog-agent/pkg/util/fargate"
+	"github.com/DataDog/datadog-agent/pkg/util/log"
 	"github.com/DataDog/datadog-agent/pkg/util/system"
 )
 
-var (
-	// overridden in tests
-	hostCPUCount = system.HostCPUCount
-)
+// overridden in tests
+var hostCPUCount = system.HostCPUCount
 
 func formatUser(fp *procutil.Process, uidProbe *LookupIdProbe) *model.ProcessUser {
 	var username string
@@ -88,4 +91,28 @@ func calculatePct(deltaProc, deltaTime, numCPU float64) float32 {
 	// Avoid reporting negative CPU percentages when this occurs
 	pct = math.Max(pct, 0.0)
 	return float32(pct)
+}
+
+// warnECSLinuxFargateMisconfig pidMode is currently not supported on fargate windows, see docs: https://docs.aws.amazon.com/AmazonECS/latest/developerguide/task_definition_parameters.html#task_definition_pidmode
+func warnECSFargateMisconfig(containers []*model.Container) {
+	if fargate.IsFargateInstance() && !isECSFargatePidModeSetToTask(containers) {
+		log.Warn(`Process collection is likely misconfigured. Please ensure your task definition has "pidMode":"task" set. See https://docs.datadoghq.com/integrations/ecs_fargate/?tab=webui#process-collection for more information.`)
+	}
+}
+
+// isECSLinuxFargatePidModeSetToTask checks if pidMode is set to task in task (which will prevent process collection) definition using specific container existence
+func isECSFargatePidModeSetToTask(containers []*model.Container) bool {
+	if !fargate.IsFargateInstance() {
+		return false
+	}
+
+	// aws-fargate-pause container only exists when "pidMode"" is set to "task" on ecs fargate
+	ecsContainerNameTag := fmt.Sprintf("%s:%s", tags.EcsContainerName, "aws-fargate-pause")
+	for _, c := range containers {
+		// container fields are not yet populated with information from tags at this point, so we need to check the tags
+		if slices.Contains(c.Tags, ecsContainerNameTag) {
+			return true
+		}
+	}
+	return false
 }
