+++ pkg/process/checks/process_nix_test.go
@@ -19,6 +19,8 @@ import (
 	model "github.com/DataDog/agent-payload/v5/process"
 	"github.com/shirou/gopsutil/v4/cpu"
 
+	"github.com/DataDog/datadog-agent/comp/core/tagger/tags"
+	"github.com/DataDog/datadog-agent/pkg/config/env"
 	"github.com/DataDog/datadog-agent/pkg/process/metadata/parser"
 	"github.com/DataDog/datadog-agent/pkg/process/procutil"
 )
@@ -375,6 +377,7 @@ func TestFormatCPUTimes(t *testing.T) {
 		})
 	}
 }
+
 func TestProcessGPUTagging(t *testing.T) {
 	p := []*procutil.Process{
 		makeProcess(1, "git clone google.com"),
@@ -424,6 +427,81 @@ func TestProcessGPUTagging(t *testing.T) {
 			for _, proc := range procs[""] {
 				assert.Equal(t, proc.Tags, tc.pidToGPUTags[proc.Pid])
 			}
+		})
+	}
+}
+
+func TestProcessIsECSFargatePidModeSetToTaskLinux(t *testing.T) {
+	for _, tc := range []struct {
+		description    string
+		containers     []*model.Container
+		fargateEnabled bool
+		expected       bool
+	}{
+		{
+			description: "ecs linux fargate pidMode not set to task",
+			containers: []*model.Container{
+				{
+					Tags: []string{
+						fmt.Sprintf("%s:%s", tags.EcsContainerName, "other container"),
+						fmt.Sprintf("%s:%s", tags.ContainerName, "aws-fargate-pause"),
+						fmt.Sprintf("%s:%s", tags.ImageName, "aws-fargate-pause"),
+					},
+				},
+			},
+			fargateEnabled: true,
+			expected:       false,
+		},
+		{
+			description: "ecs linux fargate pidMode set to task",
+			containers: []*model.Container{
+				{
+					Tags: []string{
+						fmt.Sprintf("%s:%s", tags.EcsContainerName, "aws-fargate-pause"),
+						fmt.Sprintf("%s:%s", tags.ContainerName, "some container name"),
+						fmt.Sprintf("%s:%s", tags.ImageName, "some image name"),
+					},
+				},
+			},
+			fargateEnabled: true,
+			expected:       true,
+		},
+		{
+			description: "ecs linux fargate pidMode task container exists but not on fargate",
+			containers: []*model.Container{
+				{
+					Tags: []string{
+						fmt.Sprintf("%s:%s", tags.EcsContainerName, "aws-fargate-pause"),
+						fmt.Sprintf("%s:%s", tags.ContainerName, "some container name"),
+						fmt.Sprintf("%s:%s", tags.ImageName, "some image name"),
+					},
+				},
+			},
+			fargateEnabled: false,
+			expected:       false,
+		},
+		{
+			description: "ecs linux fargate pidMode task container does not exist and not on fargate",
+			containers: []*model.Container{
+				{
+					Tags: []string{
+						fmt.Sprintf("%s:%s", tags.EcsContainerName, "some container name"),
+						fmt.Sprintf("%s:%s", tags.ContainerName, "some container name"),
+						fmt.Sprintf("%s:%s", tags.ImageName, "some image name"),
+					},
+				},
+			},
+			fargateEnabled: false,
+			expected:       false,
+		},
+	} {
+		t.Run(tc.description, func(t *testing.T) {
+			if tc.fargateEnabled {
+				env.SetFeatures(t, env.ECSFargate)
+			} else {
+				env.ClearFeatures()
+			}
+			assert.Equal(t, tc.expected, isECSFargatePidModeSetToTask(tc.containers))
 		})
 	}
 }
