+++ pkg/process/checks/process_windows.go
@@ -18,10 +18,8 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/process/procutil"
 )
 
-var (
-	// overridden in tests
-	numCPU = runtime.NumCPU
-)
+// overridden in tests
+var numCPU = runtime.NumCPU
 
 func formatUser(fp *procutil.Process, _ *LookupIdProbe) *model.ProcessUser {
 	return &model.ProcessUser{
@@ -66,4 +64,11 @@ func calculatePct(deltaProc, deltaTime, numCPU float64) float32 {
 	// Avoid reporting negative CPU percentages when this occurs
 	overalPct = math.Max(overalPct, 0.0)
 	return float32(overalPct)
+}
+
+// ecs fargate checks only support linux
+func warnECSFargateMisconfig(containers []*model.Container) {}
+
+func isECSFargatePidModeSetToTask(containers []*model.Container) bool {
+	return false
 }
