+++ releasenotes/notes/ecs-fargate-pidmode-misconfiguration-f7887b42ad48af54.yaml
@@ -1,4 +1,4 @@
 ---
 enhancements:
   - |
-    ECS Fargate (Linux): added a one-time warning when `pidMode` is not set to `task`, which may prevent process collection.
+    ECS Fargate (Linux): added a one-time warning when `pidMode` is not set to `task`, which will prevent process collection.
